refactor game renderer to game loop
