+++ TheAlienThatEatsTheCarrot/GameRenderer/GameRenderer.swift
@@ -2,37 +2,16 @@
 //  GameRenderer.swift
 //  TheAlienThatEatsTheCarrot
 //
-//  Created by Justin Cheah Yun Fei on 15/3/24.
+//  Created by Justin Cheah Yun Fei on 22/3/24.
 //
 
 import Foundation
-import QuartzCore
 
 class GameRenderer {
-    private var displayLink: CADisplayLink?
-    var gameEngine: GameEngine?
+    let camera: Camera
 
-    func start() {
-        self.displayLink = createDisplayLink()
+    init(camera: Camera) {
+        self.camera = camera
     }
 
-    func stop() {
-        displayLink?.remove(from: .main, forMode: .default)
-        displayLink?.invalidate()
-    }
-
-    func createDisplayLink() -> CADisplayLink {
-        let displaylink = CADisplayLink(target: self,
-                                        selector: #selector(step))
-        displaylink.add(to: .current, forMode: .default)
-        return displaylink
-    }
-
-    @objc func step(displaylink: CADisplayLink) {
-        if gameEngine == nil {
-            return
-        }
-        let timeInterval = displaylink.targetTimestamp - displaylink.timestamp
-        gameEngine?.update(deltaTime: timeInterval)
-    }
 }
